requirements.txt
bigquery==0.0.15
fastapi==0.68.1
furl==2.1.2
google-cloud-bigquery==2.23.2
pydantic==1.8.2
pytest==6.2.4 # including here so we can run tests in the container
python-dotenv==0.19.1
uvicorn==0.15.0
